% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{clean_lr_file}
\alias{clean_lr_file}
\title{Sources the \code{data-preparation.sh} script, which cleans the lr-file for commas}
\usage{
clean_lr_file()
}
\description{
Sources the \code{data-preparation.sh} script, which cleans the lr-file for commas
}
