% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/06-download.R
\name{hopi_get}
\alias{hopi_get}
\title{Get House Price Data}
\usage{
hopi_get(
  release = last_release(),
  frequency = "monthly",
  classification = "nuts1"
)
}
\arguments{
\item{release}{The release to download the data from. The index is begin updated
quarterly and is subject to revisions. Defaults at last release.}

\item{frequency}{The frequency to download the data. Can be \code{annual}, \code{quarterly} or \code{monthly}.}

\item{classification}{The regional classification to download the data from.
Can be one of  \code{aggregate}, \code{nuts1}, \code{nuts2}, \code{nuts3}.}
}
\value{
A tibble with the first column to be Date, and the rest to be house Prices.
}
\description{
The Housing Observatory Price Index (hopi) captures changes in the value of
residential properties with the repeated sales index. Data are available at
national and regional levels(nuts1, nuts2, nuts3).
}
\examples{
\donttest{
hopi_get(classification = "aggregate", frequency = "quarterly")
}
}
