% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-process.R
\docType{data}
\name{lr_url}
\alias{lr_url}
\title{Land registry transcaction data file}
\format{
An object of class \code{character} of length 1.
}
\usage{
lr_url
}
\description{
Land registry transcaction data file
}
\keyword{datasets}
