% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/releases.R
\name{release_dates}
\alias{release_dates}
\alias{avail_releases}
\alias{last_release}
\alias{next_release}
\alias{next_release_to_date}
\title{Get the release dates for the hopi}
\usage{
release_dates()

avail_releases()

last_release()

next_release()

next_release_to_date()
}
\description{
Get the release dates for the hopi
}
