% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-rsindex.R
\name{rsindex}
\alias{rsindex}
\title{Calculate the repeated-sales index}
\usage{
rsindex(
  data,
  gclass = c("nuts1", "nuts2", "nuts3", "countries", "uk", "london_effect"),
  freq = c("monthly", "quarterly", "annual", "daily", "weekly"),
  period_trans = 100,
  ntras_low = 1,
  ntrans_high = 8,
  abs_annual_ret = 0.15
)
}
\description{
Calculate the repeated-sales index
}
