% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{tidy_lr_file}
\alias{tidy_lr_file}
\title{Sources the \code{data-preparation.sh} script, which cleans the raw-file for commas}
\usage{
tidy_lr_file()
}
\description{
Sources the \code{data-preparation.sh} script, which cleans the raw-file for commas
}
