% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update.R
\name{update}
\alias{update}
\alias{update_monthly}
\alias{update_quarterly}
\alias{update_annual}
\title{Update a version}
\usage{
update(td, release_name = next_release())

update_monthly(td, release_name = next_release())

update_quarterly(td, release_name = next_release())

update_annual(td, release_name = next_release())
}
\arguments{
\item{td}{the output of \code{process_data()}.}

\item{release_name}{how to name the release, defaults to \code{next_release()}.}
}
\description{
Update a version
}
