% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00-preprocess.R
\name{write_nuts}
\alias{write_nuts}
\title{Write the nuts classification into an Rds file}
\usage{
write_nuts(path = "nuts.rds")
}
\arguments{
\item{path}{the path to write the rds file.}
}
\description{
Write the nuts classification into an Rds file
}
