% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-rsindex.R, R/rsindex.R
\name{rsindex}
\alias{rsindex}
\title{Calculate the repeated-sales index}
\usage{
rsindex(
  data,
  gclass = c("nuts1", "nuts2", "nuts3", "countries", "uk", "london_effect"),
  freq = c("monthly", "quarterly", "annual", "daily", "weekly"),
  period_trans = 100,
  ntras_low = 1,
  ntrans_high = 8,
  abs_annual_ret = 0.15
)

rsindex(
  data,
  gclass = c("nuts1", "nuts2", "nuts3", "countries", "uk", "london_effect"),
  freq = c("monthly", "quarterly", "annual", "daily", "weekly"),
  period_trans = 100,
  ntras_low = 1,
  ntrans_high = 8,
  abs_annual_ret = 0.15
)
}
\arguments{
\item{data}{The data to use}

\item{gclass}{geographical classification}

\item{freq}{frequenct}

\item{period_trans}{minimum period of translations}

\item{ntras_low}{minimum number of translations}

\item{ntrans_high}{maximum number of translations}

\item{abs_annual_ret}{should not exceed the absolute annual return}
}
\description{
Calculate the repeated-sales index

Calculate the House Price Index using the Repeated Sales methodology
}
